[Quantity System Framework 1.1.1]
- Whole optimization for the library. in all 4 basic operations
- Parsing optmization in the DLR part.
- factorial operator added '!' {calculation only occur for the integer part}


- fixed the exponent in metric units  10<m>+100<are>^0.5  produced error in calculations.

- Volume Flow Rate  quantity added
- cfm unit added  (cubic feet per minute)
- gpm unit added

+++ QuantitySystemSolution/QuantitySystemTestingProject/LengthDescriptorTest.cs
@@ -68,12 +68,12 @@ namespace QuantitySystemTestingProject
         [TestMethod()]
         public void LengthDescriptorConstructorTest()
         {
-            int normalExponent = 1; // TODO: Initialize to an appropriate value
-            int radiusExponent = -1; // TODO: Initialize to an appropriate value
+            float normalExponent = 1; // TODO: Initialize to an appropriate value
+            float radiusExponent = -1; // TODO: Initialize to an appropriate value
 
             LengthDescriptor target = new LengthDescriptor(normalExponent, radiusExponent);
 
-            Assert.AreEqual<int>(0, target.Exponent);
+            Assert.AreEqual<float>(0, target.Exponent);
         }
 
         /// <summary>
